서비스 코드분리/ dbmate 미적용
