+++ donghoon/.gitignore
@@ -1,0 +1,2 @@
+.DS_STORE
+node_modules
